requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pymc>=5.0.0
hmmlearn>=0.3.0
pykalman>=0.9.5
statsmodels>=0.14.0
cvxpy>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
pydantic>=2.0.0
pyyaml>=6.0
scipy>=1.11.0
